Struct patricia_trie::FatDBMut

pub struct FatDBMut<'db, H, C> where
    H: Hasher + 'db,
    C: NodeCodec<H>,  { /* fields omitted */ }

A mutable Trie implementation which hashes keys and uses a generic HashDB backing database. Additionaly it stores inserted hash-key mappings for later retrieval.

Use it as a Trie or TrieMut trait object.

Methods

impl<'db, H, C> FatDBMut<'db, H, C> where
H: Hasher,
C: NodeCodec<H>,

pub fn new(db: &'db mut HashDB<H>, root: &'db mut H::Out) -> Self

Create a new trie with the backing database db and empty root Initialise to the state entailed by the genesis block. This guarantees the trie is built correctly.

pub fn from_existing(
db: &'db mut HashDB<H>,
root: &'db mut H::Out
) -> Result<Self, H::Out, C::Error>

Create a new trie with the backing database db and root.

Returns an error if root does not exist.

pub fn db(&self) -> &HashDB<H>

Get the backing database.

pub fn db_mut(&mut self) -> &mut HashDB<H>

Get the backing database.