Struct patricia_trie::Lookup

pub struct Lookup<'a, H: Hasher + 'a, C: NodeCodec<H>, Q: Query<H>> {
    pub db: &'a HashDB<H>,
    pub query: Q,
    pub hash: H::Out,
    pub marker: PhantomData<C>,
}

Trie lookup helper object.

Fields
db: &'a HashDB<H>

database to query from.

query: Q

Query object to record nodes and transform data.

hash: H::Out

Hash to start at

marker: PhantomData<C>
Methods

impl<'a, H, C, Q> Lookup<'a, H, C, Q> where
H: Hasher + 'a,
C: NodeCodec<H> + 'a,
Q: Query<H>,

pub fn look_up(
self,
key: NibbleSlice
) -> Result<Option<Q::Item>, H::Out, C::Error>

Look up the given key. If the value is found, it will be passed to the given function to decode or copy.