Trait patricia_trie::TrieIterator

pub trait TrieIterator<H: Hasher, C: NodeCodec<H>>: Iterator {
    fn seek(
        &mut self, 
        key: &[u8]
    ) -> Result<(), H::Out, <C as NodeCodec<H>>::Error>;
}

A trie iterator that also supports random access (seek()).

Required Methods

fn seek(&mut self, key: &[u8]) -> Result<(), H::Out, <C as NodeCodec<H>>::Error>

Position the iterator on the first element with key > key