Trait patricia_trie::TrieMut

pub trait TrieMut<H: Hasher, C: NodeCodec<H>> {
    fn root(&mut self) -> &H::Out;

    fn is_empty(&self) -> bool;

    fn get<'a, 'key>(
        &'a self, 
        key: &'key [u8]
    ) -> Result<Option<DBValue>, H::Out, C::Error>
    where
        'a: 'key;

    fn insert(
        &mut self, 
        key: &[u8], 
        value: &[u8]
    ) -> Result<Option<DBValue>, H::Out, C::Error>;

    fn remove(
        &mut self, 
        key: &[u8]
    ) -> Result<Option<DBValue>, H::Out, C::Error>;

    fn contains(&self, key: &[u8]) -> Result<bool, H::Out, C::Error> { ... }
}

A key-value datastore implemented as a database-backed modified Merkle tree.

Required methods

fn root(&mut self) -> &H::Out

Return the root of the trie.

fn is_empty(&self) -> bool

Is the trie empty?

fn get<'a, 'key>(
&'a self,
key: &'key [u8]
) -> Result<Option<DBValue>, H::Out, C::Error> where
'a: 'key,

What is the value of the given key in this trie?

fn insert(
&mut self,
key: &[u8],
value: &[u8]
) -> Result<Option<DBValue>, H::Out, C::Error>

Insert a key/value pair into the trie. An empty value is equivalent to removing key from the trie. Returns the old value associated with this key, if it existed.

fn remove(&mut self, key: &[u8]) -> Result<Option<DBValue>, H::Out, C::Error>

Remove a key from the trie. Equivalent to making it equal to the empty value. Returns the old value associated with this key, if it existed.

Loading content...
Provided methods

fn contains(&self, key: &[u8]) -> Result<bool, H::Out, C::Error>

Does the trie contain a given key?

Loading content...