Struct patricia_trie::triedbmut::TrieDBMut

pub struct TrieDBMut<'a, H, C> where
    H: Hasher + 'a,
    C: NodeCodec<H>,  { /* fields omitted */ }

A Trie implementation using a generic HashDB backing database.

Use it as a TrieMut trait object. You can use db() to get the backing database object. Note that changes are not committed to the database until commit is called. Querying the root or dropping the trie will commit automatically.

Example

extern crate patricia_trie as trie;
extern crate patricia_trie_ethereum as ethtrie;
extern crate hashdb;
extern crate keccak_hash;
extern crate keccak_hasher;
extern crate memorydb;
extern crate ethereum_types;

use keccak_hash::KECCAK_NULL_RLP;
use ethtrie::{TrieDBMut, trie::TrieMut};
use hashdb::DBValue;
use keccak_hasher::KeccakHasher;
use memorydb::*;
use ethereum_types::H256;

fn main() {
  let mut memdb = MemoryDB::<KeccakHasher>::new();
  let mut root = H256::new();
  let mut t = TrieDBMut::new(&mut memdb, &mut root);
  assert!(t.is_empty());
  assert_eq!(*t.root(), KECCAK_NULL_RLP);
  t.insert(b"foo", b"bar").unwrap();
  assert!(t.contains(b"foo").unwrap());
  assert_eq!(t.get(b"foo").unwrap().unwrap(), DBValue::from_slice(b"bar"));
  t.remove(b"foo").unwrap();
  assert!(!t.contains(b"foo").unwrap());
}

Methods

impl<'a, H, C> TrieDBMut<'a, H, C> where
H: Hasher,
C: NodeCodec<H>,

pub fn new(db: &'a mut HashDB<H>, root: &'a mut H::Out) -> Self

Create a new trie with backing database db and empty root.

pub fn from_existing(
db: &'a mut HashDB<H>,
root: &'a mut H::Out
) -> Result<Self, H::Out, C::Error>

Create a new trie with the backing database db and root. Returns an error ifroot` does not exist.

pub fn db(&self) -> &HashDB<H>

Get the backing database.

pub fn db_mut(&mut self) -> &mut HashDB<H>

Get the backing database mutably.

pub fn commit(&mut self)

Commit the in-memory changes to disk, freeing their storage and updating the state root.