Trait patricia_trie::Trie

pub trait Trie {
    fn root(&self) -> &H256;

    fn get_with<'a, 'key, Q: Query>(
        &'a self, 
        key: &'key [u8], 
        query: Q
    ) -> Result<Option<Q::Item>>
    where
        'a: 'key;

    fn iter<'a>(&'a self) -> Result<Box<TrieIterator<Item = TrieItem> + 'a>>;

    fn is_empty(&self) -> bool { ... }

    fn contains(&self, key: &[u8]) -> Result<bool> { ... }

    fn get<'a, 'key>(&'a self, key: &'key [u8]) -> Result<Option<DBValue>>
    where
        'a: 'key,
    { ... }
}

A key-value datastore implemented as a database-backed modified Merkle tree.

Required Methods

fn root(&self) -> &H256

Return the root of the trie.

fn get_with<'a, 'key, Q: Query>(
&'a self,
key: &'key [u8],
query: Q
) -> Result<Option<Q::Item>> where
'a: 'key,

Search for the key with the given query parameter. See the docs of the Query trait for more details.

fn iter<'a>(&'a self) -> Result<Box<TrieIterator<Item = TrieItem> + 'a>>

Returns a depth-first iterator over the elements of trie.

Provided Methods

fn is_empty(&self) -> bool

Is the trie empty?

fn contains(&self, key: &[u8]) -> Result<bool>

Does the trie contain a given key?

fn get<'a, 'key>(&'a self, key: &'key [u8]) -> Result<Option<DBValue>> where
'a: 'key,

What is the value of the given key in this trie?